requirements.txt
anthropic
yfinance
newsapi-python
finnhub-python
alpha-vantage
feedparser
ntscraper
rich
python-dotenv
pytz
